+++ examples/sveltekit/vite.config.ts
@@ -11,8 +11,9 @@ export default defineConfig({
   ssr: {
     noExternal: [
       '@contentful/experiences-svelte',
-      '@contentful/experiences-core',
+      '@contentful/experiences-sdk-core',
       '@contentful/experiences-design',
+      '@contentful/experiences-client',
     ],
   },
   server: {