+++ packages/adapter-svelte/src/WrapWithTemplate.svelte
@@ -8,7 +8,7 @@
 <script lang="ts">
   import type { Snippet } from 'svelte';
 
-  import type { PortableTemplate } from '@contentful/experiences-core';
+  import type { PortableTemplate } from '@contentful/experiences-sdk-core';
   import { applyTokenResolver, resolveDesignProperties } from '@contentful/experiences-design';
 
   import { setContentfulTemplate, setResolvedDesign } from './context.js';