+++ packages/core/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.5.4 (2026-07-24)
+
+This was a version bump only for core to align it with other projects, there were no code changes.
+
+## 0.5.3 (2026-07-21)
+
+### 🩹 Fixes
+
+- render missing-component fallback as element in RSC [AIS-316] ([#79](https://github.com/contentful/experiences/pull/79))
+
+## 0.5.2 (2026-07-20)
+
+### 🩹 Fixes
+
+- rename experiences-core to experiences-sdk-core [AIS-305] ([#76](https://github.com/contentful/experiences/pull/76))
+
 ## 0.5.1 (2026-07-17)
 
 ### 🚀 Features