+++ packages/design/CHANGELOG.md
@@ -1,3 +1,29 @@
+## 0.5.4 (2026-07-24)
+
+### 🧱 Updated Dependencies
+
+- Updated core to 0.5.4
+
+## 0.5.3 (2026-07-21)
+
+### 🩹 Fixes
+
+- render missing-component fallback as element in RSC [AIS-316] ([#79](https://github.com/contentful/experiences/pull/79))
+
+### 🧱 Updated Dependencies
+
+- Updated core to 0.5.3
+
+## 0.5.2 (2026-07-20)
+
+### 🩹 Fixes
+
+- rename experiences-core to experiences-sdk-core [AIS-305] ([#76](https://github.com/contentful/experiences/pull/76))
+
+### 🧱 Updated Dependencies
+
+- Updated core to 0.5.2
+
 ## 0.5.1 (2026-07-17)
 
 ### 🚀 Features